Add files via upload
This is a demonstration of the working prototype of the car booking app. This will be considered Design A. A different design will also be designed with different arrangements of visual elements and interaction types. Then a usability study will be done to evaluate the comfortability of the user when interacting with these two designs.